Second part of the project - Library
New Features:
        New Aggregate - Passenger (with value objects: Address, Position, Money)
        Passenger can set target and money to offer for receiving the lift
        After finishing the course passenger pays driver for the ride and his status, target adress and amount of the money to offer are reseted
        Now during the rental driver can see free passengers with distance to them, reserve them and pick them up
        Driver class now include Balance field - after giving passengers a lift he will receive a money and after finishing a rent balance will be decreased by the value of the Total field 
        Now there is also possible to manually change car position to desired values 
        After each change of car's position, its  current as well as total distance will also increase

+++ DDD.CarRentalLib/DomainModelLayer/Models/Distance.cs
@@ -22,6 +22,7 @@ namespace DDD.CarRentalLib.DomainModelLayer.Models
             Value = value;
             Unit = DefaultUnit;
         }
+
         protected override IEnumerable<object> GetEqualityComponents()
         {
             yield return Math.Round(Value,1);
@@ -74,7 +75,7 @@ namespace DDD.CarRentalLib.DomainModelLayer.Models
         }
         private static bool IsZero(double testedValue)
         {
-            return decimal.Zero.CompareTo(testedValue) == 0;
+            return decimal.Zero.CompareTo((decimal)testedValue) == 0;
             
         }
